+++ src/Broadway.Interfaces/SecondRubric.cs
@@ -4,8 +4,8 @@ namespace NuClear.Broadway.Interfaces
 {
     public sealed class SecondRubric
     {
-        public int Code { get; set; }
-        public int CategoryCode { get; set; }
+        public long Code { get; set; }
+        public long CategoryCode { get; set; }
         public bool IsDeleted { get; set; }
         public ISet<Localization> Localizations { get; set; }
     }
